requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
build==1.2.1
cachetools==5.4.0
captcha==0.6.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
docopt==0.6.2
emoji==2.12.1
entrypoints==0.4
executing==2.0.1
extra-streamlit-components==0.1.71
Faker==26.0.0
fastjsonschema==2.20.0
favicon==0.7.0
fonttools==4.53.1
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.6.1
h11==0.14.0
h2==4.1.0
h3==3.7.7
hpack==4.0.0
htbuilder==0.6.2
httpcore==1.0.5
httpx==0.27.0
hyperframe==6.0.1
idna==3.7
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.3.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
numpy==1.26.1
openai==1.14.2
opencage==2.4.0
packaging==24.1
pandas==2.1.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
-e git+https://github.com/kumiori/philoui.git@748266d8d140e148f519b9e95c20fb12f3b8f63e#egg=philoui
pickleshare==0.7.5
pillow==10.4.0
pip-autoremove==0.10.0
pip-tools==7.4.1
pipreqs==0.5.0
platformdirs==4.2.2
plotly==5.18.0
postgrest==0.16.9
prompt_toolkit==3.0.47
protobuf==5.27.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.8.0
pymdown-extensions==10.8.1
pyparsing==3.1.2
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
realtime==1.0.6
referencing==0.35.1
requests==2.31.0
rich==13.7.1
rpds-py==0.19.1
scipy==1.11.4
setuptools==71.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
st-annotated-text==4.0.1
st-supabase-connection==1.2.0
stack-data==0.6.3
storage3==0.7.7
streamlit==1.36.0
streamlit-authenticator==0.3.3
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-elements==0.1.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.5
streamlit-faker==0.0.3
streamlit-gtag @ git+https://github.com/kumiori/streamlit-gtag.git@7a04ab228f8cb7edac23d1534e7ed8a095a53299
streamlit-image-coordinates==0.1.6
streamlit-image-select==0.6.0
streamlit-keyup==0.2.0
streamlit-modal==0.1.2
streamlit-pills==0.3.0
streamlit-player==0.1.5
streamlit-selectable-image-gallery==0.0.1
streamlit-shadcn-ui==0.1.18
streamlit-survey==0.1.0
streamlit-tags==1.2.8
streamlit-timeline==0.0.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
StrEnum==0.4.15
supabase==2.6.0
supafunc==0.5.1
tenacity==8.5.0
timezonefinder==6.2.0
tinycss2==1.3.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
wheel==0.43.0
yarg==0.1.9
